View Models refactored - refactor, moved some files to Models and segrated nested Models
