+++ voltixApp/VoltixApp/Model/KeyShare.swift
@@ -7,7 +7,12 @@
 
 import Foundation
 
-struct KeyShare: Codable {
+class KeyShare: Codable {
     let pubkey: String
     let keyshare: String
+    
+    init(pubkey: String, keyshare: String) {
+        self.pubkey = pubkey
+        self.keyshare = keyshare
+    }
 }
