--- voltixApp/VoltixApp/Protocols/DecodableDefaultSource.swift
@@ -1,11 +1,0 @@
-//
-//  DecodableDefault.swift
-//  VoltixApp
-//
-
-import Foundation
-
-protocol DecodableDefaultSource {
-    associatedtype Value: Decodable
-    static var defaultValue: Value { get }
-}
