+++ voltixApp/VoltixApp/States/DecodableDefault.swift
@@ -8,3 +8,8 @@
 import Foundation
 
 enum DecodableDefault {}
+
+protocol DecodableDefaultSource {
+    associatedtype Value: Decodable
+    static var defaultValue: Value { get }
+}
